[packages/core][m]: refactor Layout into smaller components
